requirements.txt
# llm_chat_project/requirements.txt
fastapi
uvicorn[standard]
openai
anthropic
python-dotenv